refactor(server):refactor user class
- remove sendText function to a separate file
- remove sendMail function to a separate file
- remove notifyUsers function to a separate file
- remove createToken function to a separate file
- remove validate function to a separate file
- remove flattenArray function to a separate file
- refactor test files

[Finishes]
